# backend/api.yaml
openapi: 3.0.2
info:
  title: TermPlan Backend API
  description: API documentation
  version: 1.0.0

servers:
  - url: http://localhost:3000
    description: Development server

paths:
  /verify:
    get:
      summary: Verify access token
      security:
         - bearerAuth: []
      responses:
        200:
          description: Access token verified
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: Verified
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: Unauthorized
  /register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
      responses:
        201:
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 201
                  message:
                    type: string
                    example: User created
                  accessToken:
                    type: string
                    example: <token>
                  refreshToken:
                    type: string
                    example: <token>
        400:
          description: User already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 400
                  message:
                    type: string
                    example: User with same email already exists.
        500:
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnknownErr'

  /login:
    post:
      summary: Login a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        200:
          description: User logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: User logged in
                  accessToken:
                    type: string
                    example: <token>
                  refreshToken:
                    type: string
                    example: <token>
        404:
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: User not found or wrong password
        500:
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnknownErr'
  /api/events:
    get:
      summary: Get events for the authenticated user
      security:
        - bearerAuth: []
      responses:
        200:
          description: Events retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/Events'
        500:
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnknownErr'
    post:
      summary: Create a new event
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                timeStart:
                  type: string
                  format: date-time
                name:
                  type: string
                description:
                  type: string
                  nullable: true
      responses:
        201:
          description: Event created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 201
                  message:
                    type: string
                    example: Event created
        500:
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnknownErr'
    delete:
      summary: Delete an event
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                _id:
                  type: string
      responses:
        204:
          description: Event deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 204
        500:
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnknownErr'
  /api/logout:
    post:
      summary: Logout a user
      security:
        - bearerAuth: []
      parameters:
        - in: header
          name: refreshtoken
          required: true
          schema:
            type: string
          description: The refresh token
      responses:
        200:
          description: User logged out
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: User logged out
        500:
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnknownErr'
  /refreshToken:
    post:
      summary: Refresh access token
      parameters:
        - in: header
          name: refreshtoken
          required: true
          schema:
            type: string
          description: The refresh token
      responses:
        200:
          description: Access token refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  accessToken:
                    type: string
        500:
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnknownErr'

components:
  schemas:
    UserRegister:
      type: object
      properties:
        email:
          type: string
          format: email
        username:
          type: string
        password:
          type: string
          format: password

    UserLogin:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password

    Events:
      type: object
      properties:
        _id:
          type: string
          format: objectId
        timeStart:
          type: string
          format: date-time
        timeEnd:
          type: string
          format: date-time
        name:
          type: string
        description:
          type: string
        isAllDay:
          type: boolean
        userId:
          type: string
          format: objectId
    UnknownErr:
      type: object
      properties:
        status:
          type: integer
          example: 500
        message:
          type: string
          example: Unknown error

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT